Cygwin <ctype.h> related fixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1861 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_linkam_t9x_motor.c
@@ -201,9 +201,9 @@ mxd_linkam_t9x_motor_finish_record_initialization( MX_RECORD *record )
 			record->name );
 	}
 
-	if ( islower(linkam_t9x_motor->axis_name) ) {
+	if ( islower( (int) linkam_t9x_motor->axis_name) ) {
 		linkam_t9x_motor->axis_name =
-			toupper(linkam_t9x_motor->axis_name);
+			toupper( (int) linkam_t9x_motor->axis_name);
 	}
 
 	return MX_SUCCESSFUL_RESULT;
